refactor(form): modify formComponent strucure and imports
